globalstatedb: Filter with initialized=true in getConfiguration
Currently due to the lack of database constraints and application
logic, it is possible to have multiple entries in the global_state
table. As a result, the current query is not always guaranteed to
return an entry with intialized=true.

To fix this, the query should filter for initialized=true and with the
limit 1 constraint we can ensure that if a row exists for this query,
the site has already been initialized.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>