globalstatedb: Use getConfiguration in EnsureInitialized
Instead of the redundant (and incorrect) query, we should use
getConfiguration to retrieve the site initialization status. In this
case we don't care about the site config as much as we just want to
ensure if any rows were returned as a result of running the query or
not. No rows implies that the site is not initialized and thus we
return false.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>